Reduce dependency of view model to ctk
