+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_flows.py
@@ -6,9 +6,7 @@ from customtkinter import CTkButton, CTkFrame, CTkScrollbar
 
 from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.domain.flow import Flow
-from OTAnalytics.plugin_ui.customtkinter_gui.abstract_ctk_frame_flows import (
-    AbstractFrameFlows,
-)
+from OTAnalytics.plugin_ui.customtkinter_gui.abstract_ctk_frame import AbstractCTkFrame
 from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 from OTAnalytics.plugin_ui.customtkinter_gui.helpers import get_widget_position
 from OTAnalytics.plugin_ui.customtkinter_gui.treeview_template import (
@@ -17,7 +15,7 @@ from OTAnalytics.plugin_ui.customtkinter_gui.treeview_template import (
 )
 
 
-class FrameFlows(AbstractFrameFlows):
+class FrameFlows(AbstractCTkFrame):
     def __init__(
         self,
         viewmodel: ViewModel,
